NewMap should accept the incoming context
In order to encourage users not to drop the tags from the
incoming context, adjust APIs to accept the incoming context
for convience.

Fixes #110.
