+++ plugins/grpc/grpcstats/client_handler_test.go
@@ -321,7 +321,7 @@ func TestClientDefaultCollections(t *testing.T) {
 			for _, t := range rpc.tags {
 				mods = append(mods, tag.Upsert(t.k, t.v))
 			}
-			tm, err := tag.NewMap(nil, mods...)
+			tm, err := tag.NewMap(context.Background(), mods...)
 			if err != nil {
 				t.Errorf("%q: NewMap = %v", tc.label, err)
 			}
